tests/wpt/tests/css/css-overflow/overflow-auto-scrolling-with-margin-and-transform.html

<!doctype html>
<title>Should be able to interactively scroll an overflow: auto element with a margin and transform</title>
<link rel="help" href="https://github.com/servo/servo/pull/42005">
<link rel="author" title="Jo Steven Novaryo" href="mailto:steven.novaryo@gmail.com">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/dom/events/scrolling/scroll_support.js"></script>
<style>
#scroller {
  display: block;
  margin-top: 200px;
  margin-left: 200px;
  width: 50px;
  height: 50px;
  background-color: red;
  overflow: auto;
  scrollbar-width: none;
  transform: translate(-300%, -300%);
}
#item {
  position: relative;
  top: 50px;
  left: 50px;
  width: 50px;
  height: 50px;
  background-color: green;
}
</style>
<div id="scroller">
  <div id="item">
  </div>
</div>

<script>
  promise_test(async (t) => {
    await waitForCompositorCommit();
    let rect = scroller.getBoundingClientRect();
    const actions = new test_driver.Actions().scroll(rect.left + 10, rect.top + 10, 1000, 1000);
    actions.send();

    let scroll_delta_promise = new Promise((resolve) => {
      scroller.addEventListener("scroll", () => resolve([scroller.scrollLeft, scroller.scrollTop]));
    });

    let [horizontal_delta, vertical_delta] = await scroll_delta_promise;
    assert_equals(horizontal_delta, 50, "Element scrolled horizontally");
    assert_equals(vertical_delta, 50, "Element scrolled vertically");
  }, "Margin and transform should not interfere with interactive scrolling of a scroll container");
</script>
